created and tested enpoint for modifying one entry
